backend/requirements.txt
python-dotenv
flask
pymongo
dnspython
flask-cors

